requirements.txt
alembic==1.13.2
art==6.2
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
decorator==5.1.1
Django==5.1.1
emoji==2.12.1
et-xmlfile==1.1.0
executing==2.1.0
Faker==28.4.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
googlesearch-python==1.2.5
idna==3.7
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
lorem==0.1.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy==1.11.2
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
names==0.3.0
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pwinput==1.0.3
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
simple-term-menu==1.6.4
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.34
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
WTForms==3.1.2
